+++ MC1000/Models/CartItemViewModel.cs
@@ -8,10 +8,10 @@ namespace MC1000.Models
     public class CartItemViewModel
     {
         public int ProductId { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public string Title { get; set; }
         public string Image { get; set; }
         public int Amount { get; set; }
-        public double TotalPrice => Price * Amount;
+        public decimal TotalPrice => Price * Amount;
     }
 }
